backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
boto3==1.37.34
botocore==1.37.34
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cymem==2.0.8
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
elevenlabs==1.56.1
et-xmlfile==1.1.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jmespath==1.0.1
joblib==1.4.2
language_tool_python==2.9.3
nltk==3.9.1
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.33.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.4
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
textblob==0.19.0
toml==0.10.2
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.3
websockets==15.0.1
wrapt==1.16.0
gunicorn==21.2.0
dj-database-url==1.0.0
